requirements.txt
torch
torchvision
numpy
pillow
pymongo
requests
tqdm
ftfy
regex
git+https://github.com/openai/CLIP.git